refactor: Dto 객체가 dao 패키지에 저장된 오류 수정
